+++ src/index.ts
@@ -1,1 +1,3 @@
+export { SynorCLIConfig } from './config'
+
 export { run } from '@oclif/command'